+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the "prettier-vscode" extension will be documented in thi
 ## [next]
 
 - Update loading implicit Prettier dep from `node_modules` to only occur if explicit `package.json` dep is not found in a parent directory
+- Show a custom error message / notification in the case where `prettier.prettierPath` does not reference an instance of Prettier
 
 ## [4.5.0]